src/requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==35.0.0
Django==3.2.7
djangorestframework==3.12.4
drf-jwt==1.19.1
pycparser==2.20
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
